Initial commit.
ConsoleBufferProxy doesn't work very well.  Attempts to Peek() it's stream can result in results of -1 until a Read() is done, but Read() can block forever if it's called after the `Steam>` prompt was output.  There's also a weird error that SteamCMD started throwing out that could insert anywhere in the output stream.  I treat it as if it could be between a CR and LF that are intended, but thinking back on it that might be impossible because CRLFs are handled manually by the ConsoleBufferProxy as a vertical movement in the virtual console buffer.